datastore/requirements.txt
fastapi[standard]==0.115.4
faiss-cpu==1.9.0